refactor: package to domain and view
