+++ src/main/java/racingcar/domain/CarStatus.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class CarStatus {
