+++ src/test/java/racingcar/InputTest.java
@@ -1,10 +1,12 @@
 package racingcar;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.CarName;
+import racingcar.view.Input;
 
 import static org.assertj.core.api.Assertions.*;
 
-public class InputTest {
+class InputTest {
     @Test
     void inputTest_split() {
         assertThat(Input.split("pobi,crong,honux").stream().map(CarName::getName)).contains("pobi", "crong", "honux");
